+++ processing_options.go
@@ -116,6 +116,7 @@ type processingOptions struct {
 	Crop         cropOptions
 	Format       imageType
 	Quality      int
+	MaxBytes     int
 	Flatten      bool
 	Background   rgbColor
 	Blur         float32
@@ -193,6 +194,7 @@ func newProcessingOptions() *processingOptions {
 			Gravity:      gravityOptions{Type: gravityCenter},
 			Enlarge:      false,
 			Quality:      conf.Quality,
+			MaxBytes:     0,
 			Format:       imageTypeUnknown,
 			Background:   rgbColor{255, 255, 255},
 			Blur:         0,
@@ -330,6 +332,16 @@ func parseDimension(d *int, name, arg string) error {
 	return nil
 }
 
+func parseBoolOption(str string) bool {
+	b, err := strconv.ParseBool(str)
+
+	if err != nil {
+		logWarning("`%s` is not a valid boolean value. Treated as false", str)
+	}
+
+	return b
+}
+
 func isGravityOffcetValid(gravity gravityType, offset float64) bool {
 	if gravity == gravityCenter {
 		return true
@@ -397,7 +409,7 @@ func applyEnlargeOption(po *processingOptions, args []string) error {
 		return fmt.Errorf("Invalid enlarge arguments: %v", args)
 	}
 
-	po.Enlarge = args[0] != "0"
+	po.Enlarge = parseBoolOption(args[0])
 
 	return nil
 }
@@ -407,7 +419,7 @@ func applyExtendOption(po *processingOptions, args []string) error {
 		return fmt.Errorf("Invalid extend arguments: %v", args)
 	}
 
-	po.Extend = args[0] != "0"
+	po.Extend = parseBoolOption(args[0])
 
 	return nil
 }
@@ -527,6 +539,20 @@ func applyQualityOption(po *processingOptions, args []string) error {
 		po.Quality = q
 	} else {
 		return fmt.Errorf("Invalid quality: %s", args[0])
+	}
+
+	return nil
+}
+
+func applyMaxBytesOption(po *processingOptions, args []string) error {
+	if len(args) > 1 {
+		return fmt.Errorf("Invalid max_bytes arguments: %v", args)
+	}
+
+	if max, err := strconv.Atoi(args[0]); err == nil && max >= 0 {
+		po.MaxBytes = max
+	} else {
+		return fmt.Errorf("Invalid max_bytes: %s", args[0])
 	}
 
 	return nil
@@ -734,6 +760,8 @@ func applyProcessingOption(po *processingOptions, name string, args []string) er
 		return applyCropOption(po, args)
 	case "quality", "q":
 		return applyQualityOption(po, args)
+	case "max_bytes", "mb":
+		return applyMaxBytesOption(po, args)
 	case "background", "bg":
 		return applyBackgroundOption(po, args)
 	case "blur", "bl":