+++ CHANGELOG.md
@@ -1,6 +1,27 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- `IMGPROXY_LOG_LEVEL` config.
+- `max_bytes` processing option.
+
+### Changed
+- Docker image base is changed to Debian 10 for better stability and performance.
+
+## [2.7.0] - 2019-11-13
+### Changed
+- Boolean processing options such as `enlarge` and `extend` are properly parsed. `1`, `t`, `TRUE`, `true`, `True` are truthy, `0`, `f`, `F`, `FALSE`, `false`, `False` are falsy. All other values are treated as falsy and generate a warning message.
+
+### Fixed
+- Fix segfaults on watermarking in some cases
+
+## [2.6.2] - 2019-11-11
+### Fixed
+- Fix `format` option in presets.
+
+## [2.6.1] - 2019-10-28
+### Fixed
+- Fix loading of some GIFs by using the edge version of giflib.
 
 ## [2.6.0] - 2019-10-23
 ### Added